# Mimir query engine: avoid copying native histograms where possible